Sort monthly payments

+++ isd-backend/src/main/java/com/example/isdbackend/projection/DeliveryPrice.java
@@ -7,5 +7,5 @@ public interface DeliveryPrice {
 
     int getProviderId();
 
-    int getDeliveryPrice();
+    double getDeliveryPrice();
 }
